feat: add shared copy button
