+++ src/components/tx-result.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState } from "react";
+import { CopyButton } from "@/components/copy-button";
 
 interface Props {
   result: unknown;
@@ -14,30 +14,6 @@ function formatValue(value: unknown): string {
     value,
     (_, v) => (typeof v === "bigint" ? v.toString() : v),
     2
-  );
-}
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-
-  const handleClick = async () => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1200);
-    } catch {
-      // clipboard not available
-    }
-  };
-
-  return (
-    <button
-      type="button"
-      onClick={handleClick}
-      className="text-xs text-neutral-500 hover:text-neutral-300"
-    >
-      {copied ? "Copied" : "Copy"}
-    </button>
   );
 }
